Add labels to the data points
I had to move the code that adds the axes to the bottom so they don't
get selected by the `selectAll` call at around line 33. I'm not sure if
this is the best way to do it though.

I also increased the right side padding to accommodate the text on the
right.