Change legend vertical translate math
Instead of subtracting the amount of bottom padding plus some arbitrary
value to the graph height, I now subtracted half of the bottom padding
from the height.

With this change, the vertical translation is now mostly dependent on
the height and the bottom padding. Ok, dividing by 2 is still sort of
arbitrary, but definitely less arbitrary than adding 35.